Finished CODE SMELLS
